Tweak optimized checking if the file is relevant for the cop
