+++ lib/rubocop.rb
@@ -23,8 +23,8 @@ require_relative 'rubocop/ext/processed_source'
 
 require_relative 'rubocop/error'
 require_relative 'rubocop/file_finder'
+require_relative 'rubocop/file_patterns'
 require_relative 'rubocop/name_similarity'
-require_relative 'rubocop/optimized_patterns'
 require_relative 'rubocop/path_util'
 require_relative 'rubocop/platform'
 require_relative 'rubocop/string_interpreter'
